third_party/WebKit/LayoutTests/fast/dom/setAttributeNS-prefix-and-null-namespace-expected.txt

Test that calling setAttributeNS() with a prefixed qualifiedName and null NS throws NAMESPACE_ERR.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.createElement('test').setAttributeNS(null, 'foo:bar', 'baz') threw exception NamespaceError: Failed to execute 'setAttributeNS' on 'Element': '' is an invalid namespace for attributes..
PASS successfullyParsed is true

TEST COMPLETE
